requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9